+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rilaykit/core
 
+## 1.1.0
+
+### Minor Changes
+
+- Add field renderer
+
 ## 1.0.0
 
 ### Major Changes