+++ packages/workflow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rilaykit/workflow
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @rilaykit/forms@1.1.0
+  - @rilaykit/core@1.1.0
+
 ## 1.0.0
 
 ### Major Changes